Use `void*` instead of `char*` for ufat_file_read()
Use of `void*` is more common for raw buffers (for example memcpy(),
fread() or read() from standard libraries). In C++ it is also easier to
use `void*` because any pointer can be implicitly converted to `void*`,
while conversion to `char*` requires explicit cast.

This change comes with a small cost that pointer arithmetic is more
complicated. C/C++ standards forbid any arithmetic on `void*`. GCC (and
possibly other compilers) allows it as an extension - this can be
switched off with `-pedantic-errors` or `-Werror-pointer-arith` (or
possibly others).