Use `const void*` instead of `const char*` for ufat_file_write()
Use of `const void*` is more common for raw buffers (for example
memcpy(), fwrite() or write() from standard libraries). In C++ it is
also easier to use `const void*` because any pointer can be implicitly
converted to `const void*`, while conversion to `const char*` requires
explicit cast.

This change comes with a small cost that pointer arithmetic is more
complicated. C/C++ standards forbid any arithmetic on `const void*`. GCC
(and possibly other compilers) allows it as an extension - this can be
switched off with `-pedantic-errors` or `-Werror-pointer-arith` (or
possibly others).