Remove redundant `else` in ufat_cache_open()
As the `if` part has a `return` at the end, there is no point in
actually having `else` with a separate block.

N.B. The declaration of `d` was not moved to the beginning of the
function, as whole codebase is C99 anyway - earlier standards don't have
`long long` type.