Fix bug in write_bpb() related to "blocks per sector"
The variable log2_spb is wrong, as sector size must always be greater
than block size, therefore log2_spb would either be zero or negative. As
the variable is unsigned it cannot be negative, so if sector size would
not equal block size, all code using it would be seriously broken.

Replace it with lpg2_bps, which is non-negative.